double to decimal

+++ src/Service.Fees.Domain.Models/IAssetFees.cs
@@ -10,7 +10,7 @@ namespace Service.Fees.Domain.Models
 
         FeeType FeeType { get; }
         FeeSizeType FeeSizeType { get; }
-        double FeeSize { get; }
+        decimal FeeSize { get; }
         string FeeAssetId { get; }
         string AccountId { get; }
         string WalletId { get; }
@@ -24,7 +24,7 @@ namespace Service.Fees.Domain.Models
         [DataMember(Order = 3)] public OperationType OperationType { get; set; }
         [DataMember(Order = 4)] public FeeType FeeType { get; set; }
         [DataMember(Order = 5)] public FeeSizeType FeeSizeType { get; set; }
-        [DataMember(Order = 6)] public double FeeSize { get; set; }
+        [DataMember(Order = 6)] public decimal FeeSize { get; set; }
         [DataMember(Order = 7)] public string FeeAssetId { get; set; }
         [DataMember(Order = 8)] public string AccountId { get; set; }
         [DataMember(Order = 9)] public string WalletId { get; set; }
