+++ src/Service.Fees.Domain.Models/ISpotInstrumentFees.cs
@@ -9,9 +9,9 @@ namespace Service.Fees.Domain.Models
 
         FeeType FeeType { get; }
         FeeSizeType MakerFeeSizeType { get; }
-        double MakerFeeSize { get; }
+        decimal MakerFeeSize { get; }
         FeeSizeType TakerFeeSizeType { get; }
-        double TakerFeeSize { get; }
+        decimal TakerFeeSize { get; }
         string FeeAssetId { get; }
         string AccountId { get; }
         string WalletId { get; }
@@ -24,9 +24,9 @@ namespace Service.Fees.Domain.Models
         [DataMember(Order = 2)] public string SpotInstrumentId { get; set; }
         [DataMember(Order = 3)] public FeeType FeeType { get; set; }
         [DataMember(Order = 4)] public FeeSizeType MakerFeeSizeType { get; set; }
-        [DataMember(Order = 5)] public double MakerFeeSize { get; set; }
+        [DataMember(Order = 5)] public decimal MakerFeeSize { get; set; }
         [DataMember(Order = 6)] public FeeSizeType TakerFeeSizeType { get; set; }
-        [DataMember(Order = 7)] public double TakerFeeSize { get; set; }
+        [DataMember(Order = 7)] public decimal TakerFeeSize { get; set; }
         [DataMember(Order = 8)] public string FeeAssetId { get; set; }
         [DataMember(Order = 9)] public string AccountId { get; set; }
         [DataMember(Order = 10)] public string WalletId { get; set; }
